Clean up...

+++ src/main/java/com/tipplerow/jam/math/DoubleComparator.java
@@ -1,4 +1,18 @@
-
+/*
+ * Copyright (C) 2021 Scott Shaffer - All Rights Reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.tipplerow.jam.math;
 
 import java.util.Comparator;
@@ -14,6 +28,8 @@ import java.util.Comparator;
  * The value {@code Double.POSITIVE_INFINITY} is equal to itself and
  * greater than all finite values; {@code Double.NEGATIVE_INFINITY} is
  * equal to itself and less than all other values.
+ *
+ * @author Scott Shaffer
  */
 public final class DoubleComparator implements Comparator<Double> {
     private final double tolerance;
@@ -40,7 +56,7 @@ public final class DoubleComparator implements Comparator<Double> {
     public static double EPSILON = 2.2e-16;
 
     /**
-     * Default tolerance for floating-point comparision for quantities
+     * Default tolerance for floating-point comparison for quantities
      * of order one.
      */
     public static double DEFAULT_TOLERANCE = 1.0e-12;
@@ -247,6 +263,6 @@ public final class DoubleComparator implements Comparator<Double> {
     }
 
     @Override public int compare(Double x, Double y) {
-        return compare(x.doubleValue(), y.doubleValue(), tolerance);
+        return compare(x, y, tolerance);
     }
 }
